making a bar chart with d3.js and svg
