+++ package.json
@@ -20,7 +20,9 @@
   },
   "homepage": "https://github.com/codeplus254/d3.js#readme",
   "devDependencies": {
+    "css-loader": "^3.1.0",
     "html-webpack-plugin": "^3.2.0",
+    "style-loader": "^0.23.1",
     "webpack": "^4.36.1",
     "webpack-cli": "^3.3.6",
     "webpack-dev-server": "^3.7.2"
