+++ src/bar_chart/data.csv
@@ -1,0 +1,11 @@
+country,population
+China,1415046
+India,1354052
+United States,326767
+Indonesia,266795
+Brazil,210868
+Pakistan,200814
+Nigeria,195875
+Bangladesh,166368
+Russia,143965
+Mexico,130759
