+++ src/bar_chart/index.html
@@ -3,9 +3,9 @@
 <head>
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <title> <%= htmlWebpackPlugin.options.title %> </title>
+    <title> Let's get visual with d3.js  </title>
     <meta name="viewport" content="width=device-width,initial- scale=1"> 
 </head>
 <body>
-  <svg width="960" height="640" ></svg>
+  <svg width="960" height="640"></svg>
 </html>
