--- src/styles.css
@@ -1,4 +1,0 @@
-body {
-  margin: 0px;
-  overflow: hidden;
-}
